package refactoring util #4087
