+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/generators/util/ComparePathUtil.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.generators;
+package io.swagger.codegen.generators.util;
 
 import java.io.File;
 import java.io.IOException;
